+++ src/client/kotlin/tech/thatgravyboat/skycubed/features/overlays/map/MinimapOverlay.kt
@@ -47,8 +47,8 @@ object MinimapOverlay : Overlay {
                 backgroundBox,
                 Displays.center(90, 90, Displays.renderable(MapsWidget(
                     it,
-                    GettingState.of { (McPlayer.self!!.position().x + Maps.getCurrentOffset().x.toDouble()).toFloat() },
-                    GettingState.of { (McPlayer.self!!.position().z + Maps.getCurrentOffset().z.toDouble()).toFloat() },
+                    GettingState.of { McPlayer.self!!.position().x + Maps.getCurrentOffset().x },
+                    GettingState.of { McPlayer.self!!.position().z + Maps.getCurrentOffset().z },
                     State.of(1f),
                     { false },
                     86,
